requirements.txt
requests>=2.31
beautifulsoup4>=4.12
PyJWT[crypto]>=2.8
